Replace YAML backend with MultiJson
The crack JSON parser is based on an outdated Rails parser, using YAML
conversion and a YAML backend to parse input. This is insecure in
certain common scenarios:

https://groups.google.com/forum/#!topic/rubyonrails-security/1h2DR63ViGo

This patch replaces the YAML backend with MultiJson, and emulates as
much of the specified behaviour as possible. Many of the tests describe
successful parsing of invalid JSON into Ruby objects, however, and is
not repeatable using an actual JSON parser. As a result, several tests
currently fail.